fix: deduplicate VS setup
