+++ test/units/staking/ValidatorSharePol.test.js
@@ -1,79 +1,25 @@
 import {
-  PolygonMigration,
   TestToken,
-  ERC20Permit,
   ValidatorShare,
   StakingInfo,
   EventsHub
 } from '../../helpers/artifacts.js'
 import testHelpers from '@openzeppelin/test-helpers'
 import { checkPoint, assertBigNumberEquality, assertInTransaction } from '../../helpers/utils.js'
-import { wallets, freshDeploy, approveAndStake } from './deployment.js'
-import { buyVoucherPOL, sellVoucherPOL, sellVoucherNewPOL } from './ValidatorShareHelper.js'
+import { wallets } from './deployment.js'
+import { buyVoucherPOL, sellVoucherPOL, sellVoucherNewPOL, doDeployPOL, ValidatorDefaultStake, Dynasty } from './ValidatorShareHelper.js'
+
 const BN = testHelpers.BN
 const expectRevert = testHelpers.expectRevert
 const toWei = web3.utils.toWei
 const ZeroAddr = '0x0000000000000000000000000000000000000000'
 const ExchangeRatePrecision = new BN('100000000000000000000000000000')
-const Dynasty = 8
-const ValidatorDefaultStake = new BN(toWei('100'))
 
 describe('ValidatorSharePOL', function () {
   const wei100 = toWei('100')
 
-  async function doDeploy() {
-    await freshDeploy.call(this)
-
-    this.validatorId = '8'
-    this.validatorUser = wallets[0]
-    this.stakeAmount = ValidatorDefaultStake
-
-    await this.governance.update(
-      this.stakeManager.address,
-      this.stakeManager.interface.encodeFunctionData('updateDynastyValue', [Dynasty])
-    )
-    await this.governance.update(
-      this.stakeManager.address,
-      this.stakeManager.interface.encodeFunctionData('updateValidatorThreshold', [8])
-    )
-
-    // we need to increase validator id beyond foundation id, repeat 7 times
-    for (let i = 0; i < 7; ++i) {
-      await approveAndStake.call(this, {
-        wallet: this.validatorUser,
-        stakeAmount: this.stakeAmount,
-        acceptDelegation: true,
-        pol: true
-      })
-
-      await this.governance.update(
-        this.stakeManager.address,
-        this.stakeManager.interface.encodeFunctionData('forceUnstakePOL', [i + 1])
-      )
-      await this.stakeManager.forceFinalizeCommit()
-      await this.stakeManager.advanceEpoch(Dynasty)
-      const stakeManagerValidator = this.stakeManager.connect(
-        this.stakeManager.provider.getSigner(this.validatorUser.getChecksumAddressString())
-      )
-      await stakeManagerValidator.unstakeClaimPOL(i + 1)
-      await this.stakeManager.resetSignerUsed(this.validatorUser.getChecksumAddressString())
-    }
-
-    await approveAndStake.call(this, {
-      wallet: this.validatorUser,
-      stakeAmount: this.stakeAmount,
-      acceptDelegation: true,
-      pol: true
-    })
-    await this.stakeManager.forceFinalizeCommit()
-
-    let validator = await this.stakeManager.validators(this.validatorId)
-    this.validatorContract = await ValidatorShare.attach(validator.contractAddress)
-  }
-
-
   function deployAliceAndBob() {
-    before(doDeploy)
+    before(doDeployPOL)
     before('Alice & Bob', async function () {
       this.alice = wallets[2].getChecksumAddressString()
       this.bob = wallets[3].getChecksumAddressString()
@@ -465,7 +411,7 @@ describe('ValidatorSharePOL', function () {
 
   describe('exchangeRate', function () {
     describe('when Alice purchases voucher 2 times, 1 epoch between', function () {
-      before(doDeploy)
+      before(doDeployPOL)
 
       before(async function () {
         this.user = wallets[2].getAddressString()
@@ -502,7 +448,7 @@ describe('ValidatorSharePOL', function () {
     })
 
     describe('when Alice purchases voucher and sells it', function () {
-      before(doDeploy)
+      before(doDeployPOL)
       before(async function () {
         this.user = wallets[2].getAddressString()
         await this.polToken.mint(this.user, toWei('250'))
@@ -533,8 +479,8 @@ describe('ValidatorSharePOL', function () {
     const Alice = wallets[2].getChecksumAddressString()
     const Bob = wallets[1].getChecksumAddressString()
 
-    async function doDeployAndBuyVoucherForAliceAndBob(includeBob = false) {
-      await doDeploy.call(this)
+    async function doDeployPOLAndBuyVoucherForAliceAndBob(includeBob = false) {
+      await doDeployPOL.call(this)
 
       const stake = async ({ user, stake }) => {
         await this.polToken.mint(user, stake)
@@ -658,7 +604,7 @@ describe('ValidatorSharePOL', function () {
     }
 
     describe('when Alice sells voucher', function () {
-      before(doDeployAndBuyVoucherForAliceAndBob)
+      before(doDeployPOLAndBuyVoucherForAliceAndBob)
 
       testsellVoucherPOL({
         returnedStake: aliceStake,
@@ -673,7 +619,7 @@ describe('ValidatorSharePOL', function () {
     })
 
     describe('when delegation is disabled after voucher was purchased by Alice', function () {
-      before(doDeployAndBuyVoucherForAliceAndBob)
+      before(doDeployPOLAndBuyVoucherForAliceAndBob)
       before('disable delegation', async function () {
         await this.governance.update(
           this.stakeManager.address,
@@ -694,7 +640,7 @@ describe('ValidatorSharePOL', function () {
     })
 
     describe('when Alice sells with claimAmount greater than expected', function () {
-      before(doDeployAndBuyVoucherForAliceAndBob)
+      before(doDeployPOLAndBuyVoucherForAliceAndBob)
 
       it('reverts', async function () {
         const maxShares = await this.validatorContract.balanceOf(Alice)
@@ -704,7 +650,7 @@ describe('ValidatorSharePOL', function () {
     })
 
     describe('when locked', function () {
-      before(doDeployAndBuyVoucherForAliceAndBob)
+      before(doDeployPOLAndBuyVoucherForAliceAndBob)
 
       before(async function () {
         await this.stakeManager.testLockShareContract(this.validatorId, true)
@@ -723,7 +669,7 @@ describe('ValidatorSharePOL', function () {
     })
 
     describe('when validator unstaked', function () {
-      before(doDeployAndBuyVoucherForAliceAndBob)
+      before(doDeployPOLAndBuyVoucherForAliceAndBob)
       before(async function () {
         const stakeManagerValidator = this.stakeManager.connect(
           this.stakeManager.provider.getSigner(this.validatorUser.getChecksumAddressString())
@@ -746,7 +692,7 @@ describe('ValidatorSharePOL', function () {
 
     describe('when Alice and Bob sell within withdrawal delay', function () {
       before(async function () {
-        await doDeployAndBuyVoucherForAliceAndBob.call(this, true)
+        await doDeployPOLAndBuyVoucherForAliceAndBob.call(this, true)
       })
 
       describe('when Alice sells', function () {
@@ -779,7 +725,7 @@ describe('ValidatorSharePOL', function () {
     describe('partial sell', function () {
       describe('new API', function () {
         describe('when Alice sells', function () {
-          before(doDeployAndBuyVoucherForAliceAndBob)
+          before(doDeployPOLAndBuyVoucherForAliceAndBob)
 
           const halfStake = aliceStake.div(new BN('2'))
 
@@ -821,7 +767,7 @@ describe('ValidatorSharePOL', function () {
 
       describe('old API', function () {
         describe('when Alice sells', function () {
-          before(doDeployAndBuyVoucherForAliceAndBob)
+          before(doDeployPOLAndBuyVoucherForAliceAndBob)
 
           const halfStake = aliceStake.div(new BN('2'))
 
@@ -986,7 +932,7 @@ describe('ValidatorSharePOL', function () {
     }
 
     function runWithdrawRewardsTest(timeline) {
-      before(doDeploy)
+      before(doDeployPOL)
       before(async function () {
         delegators = {}
         totalInitialBalance = new BN(0)
@@ -1053,7 +999,7 @@ describe('ValidatorSharePOL', function () {
     })
 
     describe('when not enough rewards', function () {
-      before(doDeploy)
+      before(doDeployPOL)
 
       it('reverts', async function () {
         const validatorContractAlice = this.validatorContract.connect(this.validatorContract.provider.getSigner(Alice))
@@ -1071,7 +1017,7 @@ describe('ValidatorSharePOL', function () {
     })
 
     describe('when locked', function () {
-      before(doDeploy)
+      before(doDeployPOL)
 
       before(async function () {
         const amount = toWei('100')
@@ -1097,7 +1043,7 @@ describe('ValidatorSharePOL', function () {
     })
 
     describe('when validator unstaked', function () {
-      before(doDeploy)
+      before(doDeployPOL)
       before(async function () {
         const amount = toWei('100')
         await this.polToken.mint(Alice, amount)
@@ -1128,7 +1074,7 @@ describe('ValidatorSharePOL', function () {
 
   describe('restake', function () {
     function prepareForTest({ skipCheckpoint } = {}) {
-      before(doDeploy)
+      before(doDeployPOL)
       before(async function () {
         this.user = wallets[2].getChecksumAddressString()
 
@@ -1240,7 +1186,7 @@ describe('ValidatorSharePOL', function () {
 
   describe('unstakeClaimTokens', function () {
     function prepareForTest({ skipSell, skipBuy } = {}) {
-      before(doDeploy)
+      before(doDeployPOL)
       before(async function () {
         this.user = wallets[2].getChecksumAddressString()
